requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1
cffi==1.14.0
cryptography==2.9.2
django-filter==2.3.0
django-rest-knox==4.1.0
django==3.2.6
djangorestframework==3.12.4
importlib-metadata==1.7.0
markdown==3.2.2
pillow==8.3.1
pycparser==2.20
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
waitress==2.0.0
whitenoise==5.3.0
zipp==3.1.0
